🚚 Renamed Navbar to NavbarComponent
