+++ src/components/NavbarComponent.vue
